pyproject.toml
[project]
name = "backend"
version = "1.2.3"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "httpx",
    "packaging>=25.0",
    "platformdirs",
    "py7zr",
    "pydantic",
    "pyyaml",
    "send2trash",
    "starlette",
    "tqdm",
    "uvicorn>=0.30.1",
    "watchfiles",
]
